refactor(ai-tools): extract createIfNotExists handler — phase 1 complete
