+++ nodes/Autotask/ai-tools/tool-executor-helpers.ts
@@ -1,6 +1,10 @@
 import type { IExecuteFunctions } from 'n8n-workflow';
 import { CountOperation } from '../operations/base/count-operation';
 import type { IAutotaskEntity } from '../types';
+import type { FieldMeta } from '../helpers/aiHelper';
+import { buildFieldLookup } from './filter-builder';
+import { TYPED_REFERENCE_COMPANION_FIELDS } from '../helpers/typed-reference';
+import type { OperationContractViolation } from './operation-contracts';
 
 export const DEFAULT_QUERY_LIMIT = 10;
 export const MAX_QUERY_LIMIT = 500;
@@ -38,4 +42,126 @@ export async function executeCountOperation(
 		console.debug('[executeCountOperation] count call failed:', message);
 		return null;
 	}
+}
+
+/**
+ * Build field values for create/update from params.
+ * Only includes actual entity field values, excluding control params.
+ */
+export function buildFieldValues(
+	params: Record<string, unknown>,
+	excludeKeys: string[],
+	writeFields: FieldMeta[],
+): Record<string, unknown> {
+	const result: Record<string, unknown> = {};
+	const writeFieldLookup = buildFieldLookup(writeFields);
+	const exclude = new Set<string>([
+		...excludeKeys,
+		'resource',
+		'operation',
+		'filter_field',
+		'filter_op',
+		'filter_value',
+		'filter_field_2',
+		'filter_op_2',
+		'filter_value_2',
+		'filter_logic',
+		'limit',
+		'offset',
+		'fields',
+		'recency',
+		'recency_field',
+		'since',
+		'until',
+		'domain',
+		'domainOperator',
+		'searchContactEmails',
+		'impersonationResourceId',
+		'proceedWithoutImpersonationIfDenied',
+		'dedupFields',
+		'errorOnDuplicate',
+		'updateFields',
+		'outputMode',
+		'targetOperation',
+		'filtersJson',
+		'returnAll',
+		'company',
+		// Companion fields for typed-reference resolution (ticketLookupField, projectLookupField, ...).
+		// These are schema fields consumed by the resolver, never sent to the API.
+		...Array.from(TYPED_REFERENCE_COMPANION_FIELDS),
+	]);
+	for (const [key, value] of Object.entries(params)) {
+		if (value !== undefined && value !== null && value !== '' && !exclude.has(key)) {
+			const canonicalField = writeFieldLookup.get(key.toLowerCase());
+			result[canonicalField?.id ?? key] = value;
+		}
+	}
+	return result;
+}
+
+/**
+ * Parse the 'fields' param into a selectColumns-compatible array.
+ */
+export function parseFieldsParam(fields: string | undefined): string[] {
+	if (!fields || typeof fields !== 'string') return [];
+	return fields
+		.split(',')
+		.map((f) => f.trim())
+		.filter(Boolean);
+}
+
+/**
+ * Normalise operation names to canonical forms used by the executor.
+ */
+export function normaliseOperation(operation: string): string {
+	const key = operation.trim().toLowerCase();
+	switch (key) {
+		case 'getmany':
+			return 'getMany';
+		case 'whoami':
+			return 'whoAmI';
+		case 'getposted':
+			return 'getPosted';
+		case 'getunposted':
+			return 'getUnposted';
+		case 'searchbydomain':
+			return 'searchByDomain';
+		case 'searchbyidentity':
+			return 'searchByIdentity';
+		case 'slahealthcheck':
+			return 'slaHealthCheck';
+		case 'moveconfigurationitem':
+			return 'moveConfigurationItem';
+		case 'movetocompany':
+			return 'moveToCompany';
+		case 'transferownership':
+			return 'transferOwnership';
+		case 'createifnotexists':
+			return 'createIfNotExists';
+		case 'getbyresource':
+			return 'getByResource';
+		case 'getavailableroles':
+			return 'getAvailableRoles';
+		case 'getbyyear':
+			return 'getByYear';
+		case 'describeoperation':
+			return 'describeOperation';
+		case 'describefields':
+			return 'describeFields';
+		case 'listpicklistvalues':
+			return 'listPicklistValues';
+		default:
+			return key;
+	}
+}
+
+export function buildContractViolationNextAction(
+	resource: string,
+	operation: string,
+	violations: OperationContractViolation[],
+): string {
+	return (
+		`Call autotask_${resource} with operation '${operation}' and ensure: ` +
+		violations.map((v) => v.message).join(' ')
+	);
 }
